ref(traces): Move DisabledTraceLink to components/explore
Move DisabledTraceLink from components/links/ to components/explore/
where it better fits with other explore-related components.
