+++ static/app/views/insights/common/components/samplesTable/spanSamplesTable.tsx
@@ -4,7 +4,7 @@ import {LinkButton} from '@sentry/scraps/button';
 import {Link} from '@sentry/scraps/link';
 import {Tooltip} from '@sentry/scraps/tooltip';
 
-import {DisabledTraceLink} from 'sentry/components/links/disabledTraceLink';
+import {DisabledTraceLink} from 'sentry/components/explore/disabledTraceLink';
 import type {GridColumnHeader} from 'sentry/components/tables/gridEditable';
 import {COL_WIDTH_UNDEFINED, GridEditable} from 'sentry/components/tables/gridEditable';
 import {IconProfiling} from 'sentry/icons/iconProfiling';
